requirements.txt
aiohttp
aiosignal
annotated-types
anyio
attrs
certifi
charset-normalizer
click
colorama
dataclasses-json
Deprecated
dirtyjson
distro
frozenlist
fsspec
greenlet
h11
httpcore
httpx
idna
joblib
llama-index
llama-index-embeddings-ollama
marshmallow
multidict
mypy-extensions
nest-asyncio
networkx
nltk
numpy
openai
packaging
pandas
pydantic
pydantic_core
pypdf
python-dateutil
python-dotenv
pytz
regex
requests
six
sniffio
SQLAlchemy
tenacity
tiktoken
tqdm
typing-inspect
typing_extensions
tzdata
urllib3
wrapt
yarl